TaskPals_Backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
python-dotenv==1.0.1
mysqlclient==2.2.4  # or mysql-connector-python
SQLAlchemy==2.0.32